REST_API_JeFaisTesDevoirs/requirements.txt
Django==3.2
django-debug-toolbar==3.2.1
coverage==5.5
gunicorn==20.1.0
django-cors-headers==3.7.0
djangorestframework==3.12.4
django-filter==2.4.0
importlib-metadata==4.6.3
markdown==3.3.4
zipp==3.5.0
drf_yasg==1.20.0
environ==1.0